+++ NEWS.md
@@ -1,3 +1,10 @@
+# ZooID 0.2.0
+
+* Adding functionality, including code to process individual images, a directory of images, and a whole dataset from a directory structure.
+* Also added an image dataset (both extracted and un-extracted so users can play with their own extraction code). And added functions to unpack the dataset
+
+---
+
 # ZooID 0.1.0
 
 * Initial CRAN submission.